+++ src/main/java/com/braintreegateway/MerchantRequest.java
@@ -5,8 +5,10 @@ public class MerchantRequest extends Request {
 
     public String email;
     public String countryCodeAlpha3;
+    public String companyName;
     public PayPalOnlyAccountRequest payPalOnlyAccountRequest;
     public List<String> paymentMethods;
+    public List<String> currencies;
 
     public MerchantRequest email(String email) {
         this.email = email;
@@ -18,8 +20,18 @@ public class MerchantRequest extends Request {
         return this;
     }
 
+    public MerchantRequest companyName(String companyName) {
+        this.companyName = companyName;
+        return this;
+    }
+
     public MerchantRequest paymentMethods(List<String> paymentMethods) {
         this.paymentMethods = paymentMethods;
+        return this;
+    }
+
+    public MerchantRequest currencies(List<String> currencies) {
+        this.currencies = currencies;
         return this;
     }
 
@@ -28,7 +40,9 @@ public class MerchantRequest extends Request {
         return new RequestBuilder("merchant").
             addElement("email", email).
             addElement("countryCodeAlpha3", countryCodeAlpha3).
+            addElement("companyName", companyName).
             addElement("paymentMethods", paymentMethods).
+            addElement("currencies", currencies).
             addElement("paypalAccount", payPalOnlyAccountRequest).
             toXML();
     }